// src/app/model/ProductAttribute.model.ts
import { SafeUrl } from "@angular/platform-browser";

export declare class ProductAttribute {
  id: string;
  photoUrl: string;
  photoObject: SafeUrl;
  attributeName: string;
  attributeDescription: string;
  attributePrice: number;
  attributeCategory: string;
  numberSelection: number;
  available: number;
}